#bugfix: Fix multiple issues for E011-06 implicit numeric casting
- Fixed modulo with float returning int instead of float64
  - typeof(10 % 3.0) now returns 'real' instead of 'unknown'

- Added CASE expression parsing and evaluation
  - Implemented CaseExpr AST node and parser
  - Added evalCaseExpr for CASE WHEN evaluation
  - Supports both simple and searched CASE expressions

- Fixed compareVals for mixed int/float comparison
  - ORDER BY now correctly sorts mixed int64/float64 values
  - Previously returned 0 (equal) when comparing different types

- Fixed engine_test.go incorrect type assertion
  - Test was checking result != int(30) instead of proper type assertion

All E011-06 tests now pass.

+++ internal/QE/expr.go
@@ -259,7 +259,7 @@ func (e *ExprEvaluator) mod(a, b interface{}) interface{} {
 		if bv == 0 {
 			return nil
 		}
-		return int(av) % int(bv)
+		return float64(int64(av) % int64(bv))
 	}
 
 	aIsInt := isInteger(a)
@@ -283,7 +283,6 @@ func (e *ExprEvaluator) mod(a, b interface{}) interface{} {
 		quotient.Mul(quotient, bi)
 		result := new(big.Int).Sub(ai, quotient)
 
-		// Check if result fits in int64
 		if result.IsInt64() {
 			return result.Int64()
 		}
